requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
chromadb>=0.4.18
pypdf>=3.17.0
python-docx>=0.8.11
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
tiktoken>=0.5.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.0
openai>=1.3.0